Bruno test setup (#797)
* Added Bruno code and related fixes

* Bunch of small fixes from comments

* Fix base url's

* Define env variables at collection level as documentation

* Removed now unused files

* Init and upload should be upper constrained to 2gb as it uses form-data

* Removed events integration part

* Small fixes

* Dont add bruno environments to git as may contain secrets

* correspondence_base_url -> broker_base_url

* Remove GET body

* Consistent capitalization

---------

+++ src/Altinn.Broker.API/Controllers/FileTransferController.cs
@@ -146,7 +146,7 @@ public class FileTransferController(ILogger<FileTransferController> logger) : Co
     [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [ProducesResponseType(StatusCodes.Status503ServiceUnavailable)]
     [Route("upload")]
-    [RequestFormLimits(MultipartBodyLengthLimit = long.MaxValue)]
+    [RequestFormLimits(MultipartBodyLengthLimit = int.MaxValue)]
     [Authorize(Policy = AuthorizationConstants.Sender)]
     public async Task<ActionResult> InitializeAndUpload(
         [FromForm] FileTransferInitializeAndUploadExt form,
